feat: extract balances calc to useSwapKit, create SwapAssetSelect component, use dev api endpoint
